refactor: Racer 클래스 추가 및 Car 클래스가 Racer 클래스 상속받게 수정
